+++ gui/extension/CHANGELOG.md
@@ -14,7 +14,10 @@
 - BUG#38252997 mrs.get.current_service_metadata() returns an error
 - BUG#37052409 mrs.dump.sdkServiceFiles() reports success if service does not exist
 - BUG#37575319 Incorrect output running mrs.getStatus command
+- BUG#38384637 Unexpected type alias types for spatial data types in the Typescript SDK
+- BUG#38384773 Unable to insert collections of geometry values as GEOMETRY field in the TypeScript SDK
 -
+
 
 ## Changes in 1.19.16+9.4.0